fix(RoyalRoad): add headings to chapter text and rename helpers to utils
